% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reg_repos.R
\name{add_reg_repo}
\alias{add_reg_repo}
\title{Download and register git repository containing registrations}
\usage{
add_reg_repo(url, localdir = NULL, ...)
}
\arguments{
\item{url}{Location of remote git repository. Can accept partial github
specifications of the form "<user>/<repo>".}

\item{localdir}{Full path to local checkout location of git repository. When
\code{localdir=NULL}, the default, a sensible location is chosen using the
rappdirs function.}

\item{...}{additional arguments passed to \code{git2r::clone} e.g.
credentials for private repo.}
}
\description{
Download and register git repository containing registrations
}
\examples{
\dontrun{
add_reg_repo("jefferislab/BridgingRegistrations")
add_reg_repo("jefferislab/MirrorRegistrations")
# update all current registration repositories
update_reg_repos()
}
}
\seealso{
\code{\link{update_reg_repos}}
}

