% Generated by roxygen2 (4.0.0): do not edit by hand
\name{displaySlice}
\alias{displaySlice}
\title{Display an image slice in 3D}
\usage{
displaySlice(brain, slice, ...)
}
\arguments{
\item{brain}{template brain (e.g. IS2) of the slice.}

\item{slice}{path to the slice PNG image to display.}

\item{z}{the z position of the slice. By default the slice will be positioned
midway through the brain.}

\item{...}{extra arguments to pass to persp3d.}
}
\description{
Display an image slice in 3D
}

