% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reg_repos.R
\name{download_jefferislab_registrations}
\alias{download_jefferislab_registrations}
\title{Download and install Jefferis lab registrations}
\usage{
download_jefferislab_registrations()
}
\description{
Uses \code{\link{download_reg_repo}} to download git repositories containing
bridging and mirroring registrations for Drosophila melanogaster, simulans
and virilis template brains.
}
\seealso{
\url{https://github.com/jefferislab/BridgingRegistrations}, \url{https://github.com/jefferislab/MirrorRegistrations}, and \url{https://github.com/jefferislab/DrosophilidBridgingRegistrations}.

\code{\link{download_reg_repo}}
}

