% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.r
\docType{data}
\name{Dsim.surf}
\alias{Dsim.surf}
\title{Surface model of the Drosophila simulans reference brain}
\description{
The Dsim reference brain is a shape averaged template brain generated using
the CMTK avg_adm tool. See
\url{https://github.com/jefferislab/MakeAverageBrain} for the relevant code.
}
\examples{
\dontrun{
# Depends on nat
plot3d(Dsim.surf)
}
}

