% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.r
\docType{data}
\name{FCWB}
\alias{FCWB}
\title{Template information for the FCWB FlyCircuit reference brain}
\description{
The FCWB reference brain is a shape averaged template brain generated using
the CMTK avg_adm tool. See
\url{https://github.com/jefferislab/MakeAverageBrain} for the relevant code.
}

