% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.r
\docType{data}
\name{IBN}
\alias{IBN}
\title{Template information for the Insect Brain Nomenclature reference brain}
\description{
Template used for the study "A Systematic Nomenclature for the Insect Brain"
(doi: 10.1016/j.neuron.2013.12.01).
}
\details{
Constructed from the green channel of serial sections of a Drosophila
melanogaster brain that was tricolor-labeled with a combination of the
following reporters:\itemize{
\item cytoplasmic DsRed (UAS-DsRed; Verkhusha et al., 2001; red channel),
\item presynaptic GFP (synaptic-vesicles-targeted UAS-n-syb-GFP; Estes et
al., 2000; green channel),
\item postsynaptic Rdl-hemagglutinin (GABA-receptor-targeted UAS-Rdl-HA;
Sanchez-Soriano et al., 2005; blue channel).
} These were expressed with the pan-neuronal elav-GAL4
expression driver (C155; Lin and Goodman, 1994).

Note that, because of the excess amount of ectopic protein expression, the
presynaptic and postsynaptic labelling (especially the latter) was not
completely specific, visualizing not only synaptic sites but also cell bodies
and large fiber bundles.

The image was acquired using a confocal laser-scanning microscope LSM510
(Zeiss) with a 40x water-immersion C-Apochromat objective (NA = 1.2). Each
section, of 135 total, was recorded with a resolution of 1024 by 1024 pixels
and 1.41 micron optical z-slice steps. These were then down-sampled to a
resolution of 512 by 512 pixels.
}

