% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.r
\docType{data}
\name{JFRC2013.surf}
\alias{JFRC2013.surf}
\title{Surface model of the JFRC2013 brain}
\description{
Generated in Amira from a 2 micron downsampled and median filtered version of
the JFRC2013 template brain.
}
\examples{
\dontrun{
library(nat)
plot3d(JFRC2013)
plot3d(JFRC2013.surf)
plot3d(JFRC2013.surf, col='red', alpha=0.3)
}
}
\references{
Aso, Y., Sitaraman, D., Ichinose, T., Kaun, K. R., Vogt, K., Belliart-Guerin,
 G., Placais, P.-Y., Robie, A. A., Yamagata, N., Schnaitmann, C., Rowell, W. J.,
 Johnston, R. M., Ngo, T.-T. B., Chen, N., Korff, W., Nitabach, M. N.,
 Heberlein, U., Preat, T., Branson, K. M., Tanimoto, H. and Rubin, G. M. (2014b).
 Mushroom body output neurons encode valence and guide memory-based action selection in Drosophila.
 Elife 3, e04580.
 \href{http://dx.doi.org/10.7554/eLife.04577}{doi:10.7554/eLife.04577}
}

