% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRC2018.R
\docType{data}
\name{JRC2018F}
\alias{JRC2018F}
\alias{JRC2018U}
\alias{JRC2018F.surf}
\alias{JRC2018U.surf}
\title{Template information and surface model for the JRC2018 reference brains}
\format{
An object of class \code{templatebrain} of dimension 1652 x 768 x 479.

An object of class \code{templatebrain} of dimension 1652 x 773 x 456.

An object of class \code{hxsurf} (inherits from \code{list}) of length 4.

An object of class \code{hxsurf} (inherits from \code{list}) of length 4.
}
\usage{
JRC2018F

JRC2018U

JRC2018F.surf

JRC2018U.surf
}
\description{
The \code{JRC2018F} reference brains is an average template
  brain constructed from brains labelled with brp-SNAP, dehydrated, and
  mounted in DPX and imaged at 0.19 x 0.19 x 0.38 microns. The image was
  downsampled in XY to result in a 0.38 micron isotropic voxel size, which we
  take to be the standard JRC2018F space.

The \code{JRC2018U} reference brain was constructed as for
  \code{JRC2018F} brain but pooling both male and female brains.

\code{JRC2018F.surf} and \code{JRC2018U.surf} were generated in
  Amira from a 2 micron downsampled and Lanczos filtered version of the
  JRC2018F template brain. A surface was then generated with a threshold
  level of 20; this was then smoothed and downsampled. See
  \code{data-raw/JFRC2018_FEMALE_surface.hx} or
  \code{data-raw/JRC2018_UNISEX_surface.hx} for details.
}
\details{
For the central brain, Bogovic and Saalfeld used 36 female
  individuals (72 images including left-right flips) for the female template,
  26 male individuals (52 image with left-right flips) for the male template,
  and the union of both for the unisex brain template: 62 individuals (124
  images with left-right flips).

  Downloaded from
  \url{https://www.janelia.org/open-science/jrc-2018-brain-templates}
}
\examples{
dim(JRC2018F)
voxdims(JRC2018F)
boundingbox(JRC2018F)
JRC2018F

\dontrun{
library(nat)
plot3d(JRC2018F)
plot3d(JRC2018F.surf)
plot3d(JRC2018F.surf, col='red', alpha=0.3)
}
dim(JRC2018U)
voxdims(JRC2018U)
boundingbox(JRC2018U)
JRC2018U

\dontrun{
library(nat)
plot3d(JRC2018U)
plot3d(JRC2018U.surf)
plot3d(JRC2018U.surf, col='red', alpha=0.3)
}
}
\references{
An unbiased template of the Drosophila brain and ventral nerve
  cord John A Bogovic, Hideo Otsuna, Larissa Heinrich, Masayoshi Ito,
  Jennifer Jeter, Geoffrey W Meissner, Aljoscha Nern, Jennifer Colonell, Oz
  Malkesman, Kei Ito, Stephan Saalfeld bioRxiv 376384; doi:
  \href{https://doi.org/10.1101/376384}{doi:10.1101/376384}
}
\keyword{datasets}
