% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRC2018.R
\docType{data}
\name{JRC2018F}
\alias{JRC2018F}
\alias{JRC2018F.surf}
\title{Template information and surface model for the JRC2018F reference brain}
\format{An object of class \code{templatebrain} of dimension 1652 x 768 x 479.}
\usage{
JRC2018F
}
\description{
The \code{JRC2018F} reference brain is a an average template
  brain constructed from brains labelled with brp-SNAP, dehydrated, and
  mounted in DPX and imaged at 0.19 x 0.19 x 0.38 microns. The image was
  downsampled in XY to result in a 0.38 micron isotropic voxel size, which we
  take to be the standard JRC2018F space.
}
\details{
Downloaded from
  \url{https://www.janelia.org/open-science/jrc-2018-brain-templates}
}
\examples{
dim(JRC2018F)
voxdims(JRC2018F)
boundingbox(JRC2018F)
JRC2018F

\dontrun{
library(nat)
plot3d(JRC2018F)
plot3d(JRC2018F.surf)
plot3d(JRC2018F.surf, col='red', alpha=0.3)
}
}
\references{
An unbiased template of the Drosophila brain and ventral nerve
  cord John A Bogovic, Hideo Otsuna, Larissa Heinrich, Masayoshi Ito,
  Jennifer Jeter, Geoffrey W Meissner, Aljoscha Nern, Jennifer Colonell, Oz
  Malkesman, Kei Ito, Stephan Saalfeld bioRxiv 376384; doi:
  \href{https://doi.org/10.1101/376384}{doi:10.1101/376384}
}
\keyword{datasets}
