% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRC2018.R
\docType{data}
\name{JRC2018U}
\alias{JRC2018U}
\title{Template information and surface model for the JRC2018U reference brain}
\format{An object of class \code{templatebrain} of dimension 1210 x 566 x 174.}
\usage{
JRC2018U
}
\description{
The \code{JRC2018U} reference brain is a an average template
  brain constructed from brains JRC2018F and JRC2018M.
}
\details{
Downloaded from
  \url{https://www.janelia.org/open-science/jrc-2018-brain-templates}
}
\references{
An unbiased template of the Drosophila brain and ventral nerve
  cord John A Bogovic, Hideo Otsuna, Larissa Heinrich, Masayoshi Ito,
  Jennifer Jeter, Geoffrey W Meissner, Aljoscha Nern, Jennifer Colonell, Oz
  Malkesman, Kei Ito, Stephan Saalfeld bioRxiv 376384; doi:
  \href{https://doi.org/10.1101/376384}{doi:10.1101/376384}
}
\keyword{datasets}
