% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRCFIB2018F.R
\docType{data}
\name{JRCFIB2018F}
\alias{JRCFIB2018F}
\alias{JRCFIB2018Fraw}
\title{Template information for the JRCFIB2018F aka hemibrain dataset}
\description{
The \code{JRCFIB2018F} reference brain is a stitched FIB SEM
  volume completed at HHMI Janelia Research Campus in 2018.

  The \code{JRCFIB2018Fraw} reference brain is uncalibrated (i.e. has units
  of raw pixels), whereas \code{JRCFIB2018F} has units of microns and voxel
  dimensions of 8x8x8 * 1E-3 microns (i.e. 8 nm isotropic voxels).
}
\references{
\bold{A Connectome of the Adult Drosophila Central Brain} Shan
  Xu, C; Januszewski, Michal; Lu, Zhiyuan; Takemura, Shin-Ya; Hayworth,
  Kenneth; Huang, Gary; Shinomiya, Kazunori; Maitin-Shepard, Jeremy;
  Ackerman, David; Berg, Stuart; Blakely, Tim; Bogovic, John; Clements, Jody;
  Dolafi, Tom; Hubbard, Philip; Kainmueller, Dagmar; Katz, William; Kawase,
  Takashi; Khairy, Khaled; Leavitt, Laramie; Li, Peter H; Lindsey, Larry;
  Neubarth, Nicole; Olbris, Donald J; Otsuna, Hideo; Troutman, Eric T;
  Umayam, Lowell; Zhao, Ting; Ito, Masayoshi; Goldammer, Jens; Wolff, Tanya;
  Svirskas, Robert; Schlegel, Philipp; Neace, Erika R; Knecht, Christopher J;
  Alvarado, Chelsea X; Bailey, Dennis; Ballinger, Samantha; Borycz, Jolanta
  A; Canino, Brandon; Cheatham, Natasha; Cook, Michael; Dreyer, Marisa;
  Duclos, Octave; Eubanks, Bryon; Fairbanks, Kelli; Finley, Samantha;
  Forknall, Nora; Francis, Audrey; Hopkins, Gary Patrick; Joyce, Emily M;
  Kim, Sungjin; Kirk, Nicole A; Kovalyak, Julie; Lauchie, Shirley A; Lohff,
  Alanna; Maldonado, Charli; Manley, Emily A; McLin, Sari; Mooney, Caroline;
  Ndama, Miatta; Ogundeyi, Omotara; Okeoma, Nneoma; Ordish, Christopher;
  Padilla, Nicholas; Patrick, Christopher; Paterson, Tyler; Phillips, Elliott
  E; Phillips, Emily M; Rampally, Neha; Ribeiro, Caitlin; Robertson,
  Madelaine K; Rymer, Jon Thomson; Ryan, Sean M; Sammons, Megan; Scott, Anne
  K; Scott, Ashley L; Shinomiya, Aya; Smith, Claire; Smith, Kelsey; Smith,
  Natalie L; Sobeski, Margaret A; Suleiman, Alia; Swift, Jackie; Takemura,
  Satoko; Talebi, Iris; Tarnogorska, Dorota; Tenshaw, Emily; Tokhi, Temour;
  Walsh, John J; Yang, Tansy; Horne, Jane Anne; Li, Feng; Parekh, Ruchi;
  Rivlin, Patricia K; Jayaraman, Vivek; Ito, Kei; Saalfeld, Stephan; George,
  Reed; Meinertzhagen, Ian; Rubin, Gerald M; Hess, Harald F; Scheffer, Louis
  K; Jain, Viren; Plaza, Stephen M
  \href{https://www.biorxiv.org/content/10.1101/2020.01.21.911859v1}{bioRxiv
  doi:10.1101/2020.01.21.911859}.
}
