% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRC2018.R
\docType{data}
\name{JRCVNC2018}
\alias{JRCVNC2018}
\alias{JRCVNC2018F}
\alias{JRCVNC2018F.surf}
\title{Template information and surface models for JRC2018 VNC templates}
\format{
An object of class \code{templatebrain} of dimension 660 x 1342 x 358.

An object of class \code{hxsurf} (inherits from \code{list}) of length 4.
}
\usage{
JRCVNC2018F

JRCVNC2018F.surf
}
\description{
The \code{JRCVNC2018F} reference VNC was constructed as for
  \code{\link{JRC2018F}} brain based on 36 individuals, with a resolution of
 0.4 x 0.4 x 0.4 microns.

\code{JRCVNC2018F.surf}, was generated in
  Amira from a 2 micron downsampled, Lanczos filtered version of the
  respective template brain. A surface was then generated with a threshold
  level of 2800 and then smoothed and downsampled.
}
\examples{
JRCVNC2018F
dim(JRCVNC2018F)
voxdims(JRCVNC2018F)
boundingbox(JRCVNC2018F)

\dontrun{
library(nat)
plot3d(JRCVNC2018F)
plot3d(JRCVNC2018F.surf)
plot3d(JRCVNC2018F.surf, col='red', alpha=0.3)
}
}
\references{
An unbiased template of the Drosophila brain and ventral nerve
  cord John A Bogovic, Hideo Otsuna, Larissa Heinrich, Masayoshi Ito,
  Jennifer Jeter, Geoffrey W Meissner, Aljoscha Nern, Jennifer Colonell, Oz
  Malkesman, Kei Ito, Stephan Saalfeld bioRxiv 376384; doi:
  \href{https://doi.org/10.1101/376384}{doi:10.1101/376384}
}
\seealso{
\code{\link{JRCVNC2018F}}
}
\keyword{datasets}
