\docType{package}
\name{nat.flybrains-package}
\alias{nat.flybrains}
\alias{nat.flybrains-package}
\title{Additional flybrain-specific data and tools for use with the NeuroAnatomy Toolbox (nat)}
\description{
This package provides data, such as template brain
information and registrations, along with extra functions
to extend nat towards easy use for researchers working with
fly data.
}
\section{Template brains}{
  Information on a number of template brains is provided.
}

\section{Bridging registrations}{
  A number of registrations for bridging data from one
  template to another are provided.
}

\section{Mirroring registrations}{
  A number of registrations for mirroring data from one
  side of a template brain to the other are provided.
}

\section{Helper functions}{
  Easy-to-use functions for transforming data from one
  template brain to another, displaying slices alongside 3D
  data, etc. are provided.
}
\seealso{
\code{\link[nat]{nat}}
}
\keyword{package}

