% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot3d.BrainTemplate}
\alias{plot3d.BrainTemplate}
\title{Plot a 3D surface of a template brain}
\usage{
\method{plot3d}{BrainTemplate}(x, col = "grey", alpha = 0.3, ...)
}
\arguments{
\item{x}{the template brain to plot.}

\item{col}{the color of the surface.}

\item{alpha}{the alpha value of the surface.}

\item{...}{extra arguments to pass to \code{\link[rgl]{plot3d}}.}
}
\description{
Plot a 3D surface of a template brain
}

