% Generated by roxygen2 (4.0.1): do not edit by hand
\name{templatebrain}
\alias{templatebrain}
\alias{templatebrain.default}
\title{Construct templatebrain object for an image registration template}
\usage{
templatebrain(x, ...)

\method{templatebrain}{default}(name, regName, imageFile, type, sex,
  description)
}
\arguments{
\item{x}{the object to use to construct the templatebrain.}

\item{...}{additional arguments to pass.}

\item{name}{the name of the template.}

\item{regName}{the short name to use for finding appropriate registrations.}

\item{imageFile}{path to the image that defines the template brain.}

\item{type}{one of \code{c('single brain', 'average')}, indicating whether
the template brain has been created from just one image, or is the average
of multiple images.}

\item{sex}{the sex of the template brain. For templates with
\code{type=='average'}, the possibility of \code{sex='intersex'} exists.}

\item{description}{details of the template.}
}
\value{
A list with additional class \code{templatebrain}.
}
\description{
\code{templatebrain} objects encapsulate key information for the reference
brain in an image registration. Usually this will be a standard template
brain used for many registrations.
}
\details{
\code{templatebrain} objects are only useful for transformation processes
when the \code{BoundingBox} is specified to define the physical extent of the
volume. We use the definition of the Amira 3D visualisation and analysis
software. This corresponds to the \strong{node} centers option in the
\href{http://teem.sourceforge.net/nrrd/format.html}{NRRD format}. The
bounding box can be obtained from nrrd or amiramesh format files.
}
\seealso{
\code{\link[nat]{im3d}, \link[nat]{boundingbox}}
}

