% Generated by roxygen2 (4.0.0): do not edit by hand
\name{trans}
\alias{trans}
\title{Transform 3D object from one template brain to another, optionally using a warping registration}
\usage{
trans(template1, template2, data, transform = c("warp", "affine"), ...)
}
\arguments{
\item{template1}{source template brain (e.g. IS2) that data is in.}

\item{template2}{target template brain (e.g. IS2) that data should be transformed into.}

\item{data}{the 3D object to be mirrored.}

\item{transform}{whether to use warp (default) or affine component of
registration.}

\item{...}{extra arguments to pass to \code{\link[nat]{xform}}.}
}
\description{
Transform 3D object from one template brain to another, optionally using a warping registration
}

