% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.r
\docType{data}
\name{VNCIS1}
\alias{VNCIS1}
\title{Template information for VNCIS1}
\description{
Template information for VNCIS1
}
\examples{
VNCIS1
# originally calculated thus:
\dontrun{
as.templatebrain(read.im3d(
  "/GD/projects/CommonNeuroanatomy/VNCIS1/VNCIS1.nrrd", ReadData = FALSE),
  type="Average", sex='Intersex',
  description = "See http://dx.doi.org/10.1016/j.cub.2010.07.045 for details")
}
}
\references{
Cachero S., Ostrovsky A.D., Yu J.Y., Dickson B.J., and Jefferis
 G.S.X.E. (2010). Sexual dimorphism in the fly brain. Curr Biol 20 (18),
 1589-601.
 \href{http://dx.doi.org/10.1016/j.cub.2010.07.045}{doi:10.1016/j.cub.2010.07.045}
}
\seealso{
\code{\link{VNCIS1}}
}

