% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xform.BrainTemplate}
\alias{xform}
\alias{xform.BrainTemplate}
\title{Transform 3D object from one template brain to another, optionally using a warping registration}
\usage{
\method{xform}{BrainTemplate}(x, x2, data, transform = c("warp", "affine"))
}
\arguments{
\item{x}{template brain (e.g. IS2) that data is in.}

\item{x2}{template brain (e.g. IS2) that data should be transformed into.}

\item{data}{the 3D object to be mirrored.}

\item{transform}{whether to use warp (default) or affine component of
registration.}
}
\description{
Transform 3D object from one template brain to another, optionally using a warping registration
}

