% Generated by roxygen2 (4.0.0): do not edit by hand
\name{BrainTemplate}
\alias{BrainTemplate}
\alias{BrainTemplate.default}
\title{Construct a BrainTemplate object representing a template brain for
registration}
\usage{
BrainTemplate(x, ...)

\method{BrainTemplate}{default}(name, imageFile, type, sex, description)
}
\arguments{
\item{x}{the object to use to construct the BrainTemplate.}

\item{...}{additional arguments to pass.}

\item{name}{the name of the template.}

\item{imageFile}{path to the image that defines the template brain.}

\item{type}{one of \code{c('single brain', 'average')}, indicating whether
the template brain has been created from just one image, or is the average
of multiple images.}

\item{sex}{the sex of the template brain. For templates with
\code{type=='average'}, the possibility of \code{sex='intersex'} exists.}

\item{description}{details of the template.}
}
\value{
A list with additional class \code{BrainTemplate}.
}
\description{
There should always be a \code{BoundingBox} value set which defines the
physical extent of the volume in the same manner as the Amira 3D
visualisation and analysis software. This corresponds to the \strong{node}
centers option in the
\href{http://teem.sourceforge.net/nrrd/format.html}{NRRD format}.
}
\details{
We follow Amira's convention of setting the bounding box equal to
  voxel dimension (rather than 0) for any dimension with only 1 voxel.
}

