% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.r
\docType{data}
\name{Dmel}
\alias{Dmel}
\title{Template information for the Drosophila melanogaster reference brain}
\description{
The Dmel reference brain is a shape averaged template brain generated using
the CMTK avg_adm tool. See
\url{https://github.com/jefferislab/MakeAverageBrain} for the relevant code.
}
\details{
\href{http://dx.doi.org/10.5281/zenodo.10591}{doi:10.5281/zenodo.10591}
}

