% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reg_repos.R
\name{download_jefferislab_registrations}
\alias{download_jefferislab_registrations}
\title{Download (or update) and install Jefferis lab registrations}
\usage{
download_jefferislab_registrations(repos = c("BridgingRegistrations",
  "MirrorRegistrations", "DrosophilidBridgingRegistrations"))
}
\arguments{
\item{repos}{character vector of Jefferis Lab GitHub repositories to download
  from.}
}
\description{
Uses \code{nat.templatebrains::\link{download_reg_repo}} to download git
repositories containing bridging and mirroring registrations for Drosophila
melanogaster, simulans and virilis template brains.
}
\examples{
\dontrun{
# to download just one of these repositories do e.g.:
download_jefferislab_registrations("Mirror")

# to download your own registration repositories for use do
download_reg_repo("https://github.com/myuser/myrepo")
}
}
\references{
For a full description and sample applications of these registrations,
  please see:

Combining genome-scale Drosophila 3D neuroanatomical data by bridging template brains
James D. Manton, Aaron D. Ostrovsky, Lea Goetz, Marta Costa, Torsten Rohlfing,
Gregory S. X. E. Jefferis
\href{http://dx.doi.org/10.1101/006353}{bioRxiv doi: 10.1101/006353}
}
\seealso{
\url{https://github.com/jefferislab/BridgingRegistrations},
  \url{https://github.com/jefferislab/MirrorRegistrations}, and
  \url{https://github.com/jefferislab/DrosophilidBridgingRegistrations}.

\code{\link{download_reg_repo}}
}

