% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA-Drosophilids.r
\docType{data}
\name{Dsim}
\alias{Dsim}
\alias{Dsim.surf}
\title{Template information and surface model for D. simulans reference brain}
\description{
The Dsim reference brain is a shape averaged template brain generated using
the CMTK avg_adm tool. See
\url{https://github.com/jefferislab/MakeAverageBrain} for the relevant code.

The surface model was constructed in Amira by L. Goetz and G.
  Jefferis using a simple threshold, followed by a surface simplification to
  ~ 18,000 faces.
}
\details{
\href{http://dx.doi.org/10.5281/zenodo.1059}{doi:10.5281/zenodo.1059}
}
\examples{
\dontrun{
# Depends on nat
plot3d(Dsim.surf)
}
}
\seealso{
Other Drosophilid-Brains: \code{\link{Dmel}},
  \code{\link{DsecI}}, \code{\link{Dvir}}
}
\concept{Drosophilid-Brains}
