% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.r
\docType{data}
\name{Dsim.surf}
\alias{Dsim.surf}
\title{Surface model of the Drosophila simulans reference brain}
\description{
The Dsim reference brain is a shape averaged template brain generated using
the CMTK avg_adm tool. See
\url{https://github.com/jefferislab/MakeAverageBrain} for the relevant code.
}
\details{
\href{http://dx.doi.org/10.5281/zenodo.1059}{doi:10.5281/zenodo.1059}
}
\examples{
\dontrun{
# Depends on nat
plot3d(Dsim.surf)
}
}

