% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.r
\docType{data}
\name{IS2}
\alias{IS2}
\title{Template information for the IS2 reference brain}
\description{
Template information for the IS2 reference brain
}
\references{
Cachero S., Ostrovsky A.D., Yu J.Y., Dickson B.J., and Jefferis
 G.S.X.E. (2010). Sexual dimorphism in the fly brain. Curr Biol 20 (18),
 1589-601.
\href{http://dx.doi.org/10.1016/j.cub.2010.07.045}{doi:10.1016/j.cub.2010.07.045}
}
