% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.r
\docType{data}
\name{JFRC2013}
\alias{JFRC2013}
\alias{JRC2013}
\title{Template information and surface model for the JFRC2013 reference brain}
\format{\preformatted{List of 11
 $ name       : chr "JFRC2013"
 $ regName    : chr "JFRC2013"
 $ type       : NULL
 $ sex        : chr "Female"
 $ dims       : num [1:3] 1450 725 436
 $ voxdims    : num [1:3] 0.38 0.38 0.38
 $ origin     : num [1:3] 0 0 0
 $ BoundingBox: boundingbox [1:2, 1:3] 0 551 0 275 0 ...
 $ units      : chr [1:3] "microns" "microns" "microns"
 $ description: chr "The JFRC2013 reference brain is a spatially calibrated version of the brain used in Aso et al. 2014, \\"The neuronal architectur"| __truncated__
 $ doi        : NULL
 - attr(*, "class")= chr "templatebrain"
}}
\usage{
JFRC2013
}
\description{
The JFRC2013 reference brain is a single female brain. It is a
  spatially calibrated version of the brain used in Aso et al. 2014, "The
  neuronal architecture of the mushroom body provides a logic for associative
  learning".
}
\section{Calibration}{
 The isotropic calibration of 0.38 microns is based on a
  personal communication from Yoshi Aso on 21st May 2014, where he noted:

  It is 0.38um isotropic. Due to dehydration steps with EtOH, tissues are
  about 20% smaller than when mounted in Vectashield or other glycerol based
  mounting medium.

  No attempt was made to correct for this shrinkage artefact.
}
\keyword{datasets}

