% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.r
\docType{data}
\name{JFRC2013.surf}
\alias{JFRC2013.surf}
\title{Surface model of the JFRC2013 brain}
\description{
Generated in Amira from a 2 micron downsampled and median filtered version of
the JFRC2013 template brain.
}
\examples{
\dontrun{
library(nat)
plot3d(JFRC2013)
plot3d(JFRC2013.surf)
plot3d(JFRC2013.surf, col='red', alpha=0.3)
}
}

