% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DATA.r
\docType{data}
\encoding{UTF-8}
\name{JFRC2NP.surf}
\alias{JFRC2NP.surf}
\title{Surface model of the Insect Brain Name Working Group neuropil segmentation in
JFRC2 coordinate space}
\description{
Surface model of the Insect Brain Name Working Group neuropil segmentation in
JFRC2 coordinate space
}
\examples{
\dontrun{
# Depends on nat
plot3d(JFRC2NP.surf)
}
}
\references{
Kei Ito, Kazunori Shinomiya, Masayoshi Ito, J. Douglas Armstrong,
 George Boyan, Volker Hartenstein, Steffen Harzsch, Martin Heisenberg, Uwe
 Homberg, Arnim Jenett, Haig Keshishian, Linda L. Restifo, Wolfgang
 \enc{Rössler}{Roessler}, Julie H. Simpson, Nicholas J. Strausfeld, Roland
 Strauss, Leslie B. Vosshall, Insect Brain Name Working Group (2013). A
 systematic nomenclature for the insect brain. Neuron 81 (4), 755-765. doi:
 10.1016/j.neuron.2013.12.017.
}

