% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.r
\docType{data}
\name{T1}
\alias{T1}
\alias{T1.surf}
\title{Template information and surface model for the T1 reference brain}
\description{
Template information and surface model for the T1 reference brain

The surface model was constructed in Amira in the Jefferis Lab using a simple
threshold, followed by a surface simplification.
}
\examples{
\dontrun{
# Depends on nat
plot3d(T1.surf)
}
}

