% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.r
\docType{data}
\name{VNCIS1}
\alias{VNCIS1}
\title{Template information for VNCIS1}
\description{
Template information for VNCIS1
}
\examples{
VNCIS1
# originally calculated thus:
\dontrun{
as.templatebrain(read.im3d(
  "/GD/projects/CommonNeuroanatomy/VNCIS1/VNCIS1.nrrd", ReadData = FALSE),
  type="Average", sex='Intersex',
  description = "See http://dx.doi.org/10.1016/j.cub.2010.07.045 for details")
}
}
\seealso{
\code{\link{VNCIS1}}
}

