% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DATA.r
\docType{data}
\name{VNCIS1.surf}
\alias{VNCIS1.surf}
\title{Surface model of the VNCIS1 ventral nerve cord template brain}
\description{
The VNCIS1 template is a shape-averaged, intersex template brain generated
using the CMTK avg_adm tool.
}
\references{
Cachero S., Ostrovsky A.D., Yu J.Y., Dickson B.J., and Jefferis
  G.S.X.E. (2010). Sexual dimorphism in the fly brain. Curr Biol 20 (18),
  1589-601. See also \url{http://dx.doi.org/10.5281/zenodo.13884}
}

