% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xform_brain}
\alias{xform_brain}
\title{Transform 3D object from one template brain to another, optionally using a warping registration}
\usage{
xform_brain(x, reference, sample, ...)
}
\arguments{
\item{x}{the 3D object to be mirrored.}

\item{reference}{target template brain (e.g. IS2) that data should be transformed into.}

\item{sample}{source template brain (e.g. IS2) that data is currently in.}

\item{...}{extra arguments to pass to \code{\link[nat]{xform}}.}
}
\description{
Transform 3D object from one template brain to another, optionally using a warping registration
}

